+++ mip_package_manager/__init__.py
@@ -1,4 +1,4 @@
 """mip-package-manager - pip-style package manager for MATLAB packages
 """
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'